res fixed some GCC warnings & errors in the shader plugins.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@17275 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/shader/shaderplugins/glshader_ps1/ps1_emu_nv.cpp
@@ -117,7 +117,7 @@ void csShaderGLPS1_NV::ResetState ()
 
 void csShaderGLPS1_NV::ActivateTextureShaders ()
 {
-  if(tex_program_num != ~0)
+  if (tex_program_num != (GLuint)~0)
   {
     glCallList(tex_program_num);
     return;
